Extract "Profile"-class from Permissions
