+++ src/main/java/sirius/web/security/Permissions.java
@@ -24,7 +24,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.function.Predicate;
 
@@ -52,86 +51,6 @@ public class Permissions {
     private static final String ENABLED = "enabled";
 
     private static final Log LOG = Log.get("permissions");
-
-    /**
-     * Represents a profile defined in <tt>security.profiles</tt>.
-     */
-    private static class Profile {
-        public static final String SECURITY_PROFILES = "security.profiles";
-
-        private final String name;
-        private final Set<String> permissionsToAdd;
-        private final Set<String> permissionsToRemove;
-
-        /**
-         * Create a Profile with the given name, permissions this profiles adds and permissions this profile removes.
-         *
-         * @param name                the name of the profile
-         * @param permissionsToAdd    permissions this profile adds
-         * @param permissionsToRemove permission this profiles removes
-         */
-        Profile(String name, Set<String> permissionsToAdd, Set<String> permissionsToRemove) {
-            this.name = name;
-            this.permissionsToAdd = permissionsToAdd;
-            this.permissionsToRemove = permissionsToRemove;
-        }
-
-        /**
-         * Applies the profile to the given set of permissions.
-         *
-         * @param permissions the permissions the profile should be applied to
-         */
-        protected void apply(Set<String> permissions) {
-            if (hasPermission(name, permissions::contains)) {
-                permissions.addAll(permissionsToAdd);
-                permissions.removeAll(permissionsToRemove);
-            }
-        }
-
-        /**
-         * Validates this profile and throws exception if problems exist.
-         * <p>
-         * An exception will be thrown if the profile refers to another profile applied earlier than itself.
-         */
-        protected void validate() {
-            Extension thisProfile = Sirius.getSettings().getExtension(SECURITY_PROFILES, name);
-            for (String permission : thisProfile.getContext().keySet()) {
-                Extension otherProfile = Sirius.getSettings().getExtension(SECURITY_PROFILES, permission);
-                if (otherProfile == null || otherProfile.isDefault()) {
-                    continue;
-                }
-                if (otherProfile.compareTo(thisProfile) <= 0) {
-                    throw new IllegalStateException(Strings.apply(
-                            "Profile '%s' refers to a profile which is applied earlier than itself ('%s'). "
-                            + "Therefore the profiles will not be resolved completely. Fix this by adding priorities.",
-                            thisProfile.getId(),
-                            otherProfile.getId()));
-                }
-            }
-        }
-
-        /**
-         * Compile the given extension into a {@link Profile}.
-         *
-         * @param extension the extension to compile
-         * @return the compiled {@link Profile}
-         */
-        protected static Profile compile(Extension extension) {
-            Set<String> permissionsToAdd = new HashSet<>();
-            Set<String> permissionsToRemove = new HashSet<>();
-
-            for (Map.Entry<String, Object> permission : extension.getContext().entrySet()) {
-                if (Boolean.TRUE.equals(permission.getValue())) {
-                    permissionsToAdd.add(permission.getKey());
-                }
-                if (Boolean.FALSE.equals(permission.getValue())) {
-                    permissionsToRemove.add(permission.getKey());
-                }
-            }
-
-            return new Profile(extension.getId(), permissionsToAdd, permissionsToRemove);
-        }
-    }
 
     protected static List<Profile> profilesCache;
