+++ src/test/java/sirius/web/security/PermissionsSpec.groovy
@@ -9,7 +9,6 @@
 package sirius.web.security
 
 import sirius.kernel.BaseSpecification
-import sirius.kernel.Sirius
 
 import java.util.function.Predicate
 
@@ -37,45 +36,6 @@ class PermissionsSpec extends BaseSpecification {
         roles.contains("test-A")
         and:
         !roles.contains("test-B")
-    }
-
-    def "profile cascading is invalid when cascading to lower priority"() {
-        when:
-        Permissions
-                .Profile
-                .compile(Sirius.getSettings().getExtension("security.profiles",
-                                                           "test-cascade-to-target-with-lower-priority"))
-                .validate()
-        then:
-        def e = thrown(IllegalStateException)
-        e.getMessage() == "Profile 'test-cascade-to-target-with-lower-priority' refers to a profile which is applied " +
-                "earlier than itself ('cascade-target'). Therefore the profiles will not be resolved completely. Fix " +
-                "this by adding priorities."
-    }
-
-    def "profile cascading is invalid when cascading to equal priority"() {
-        when:
-        Permissions
-                .Profile
-                .compile(Sirius.getSettings().getExtension("security.profiles",
-                                                           "test-cascade-to-target-with-equal-priority"))
-                .validate()
-        then:
-        def e = thrown(IllegalStateException)
-        e.getMessage() == "Profile 'test-cascade-to-target-with-equal-priority' refers to a profile which is applied " +
-                "earlier than itself ('cascade-target'). Therefore the profiles will not be resolved completely. Fix " +
-                "this by adding priorities."
-    }
-
-    def "profile cascading is valid when cascading to higher priority"() {
-        when:
-        Permissions
-                .Profile
-                .compile(Sirius.getSettings().getExtension("security.profiles",
-                                                           "test-cascade-to-target-with-higher-priority"))
-                .validate()
-        then:
-        noExceptionThrown()
     }
 
     def "test hasPermission"() {
